Smart Cascade
The addition of smart cascade takes the underlying test generation of
evaporative ternary (fixed # of balls per test solution, vs. fixed # of
balls for the component with the smallest vol. frac. from phase walk,
which sucked..) and applies a smart gradient search, which recognizes
that the previous move will likely still be the best move, and that
moving backwards will obviously not help. Computational efficiency,
however, only improved by ~10%.