Stable, Robust, no glitch GANGRUNS
There is some finicky shit with the size of the step used to check the
initial gradient. A really low initial step for the 50/50 LOCDERIV
check will predict tons of incorrect metastability.

The problem with the phase flip false minima was that my moves didn’t
account for the possible asymmetry of the phases. A binary system with
trading allowed has 6 moves, not 4.